Integrate user authentication with Clerk and finalize Lab 5 deployment setup

+++ backend/src/api/controller/employeeController.ts
@@ -2,6 +2,7 @@ import "reflect-metadata";
 import { Request, Response } from "express";
 
 import * as EmployeeService from "../services/employeeService";
+import * as UserService from "../services/user";
 import {
   Controller,
   Delete,
@@ -15,12 +16,18 @@ import {
 } from "routing-controllers";
 import { validateRequest } from "../middleware/validate";
 import { employeeSchema } from "../validations/roleValidation";
+import { getAuth, requireAuth } from "@clerk/express";
+import { findOrCreateUser } from "../middleware/findOrCreateUser";
 
 @Controller()
 export class EmployeeController {
   @Get("/employee")
+  // @UseBefore(requireAuth())
   async getAll(@Req() req: Request, @Res() res: Response) {
     try {
+      const auth = getAuth(req);
+      const userId = auth.userId;
+      console.log("🚀 ~ EmployeeController ~ getAll ~ userId:", userId)
       const employees = await EmployeeService.fetchAllEmployee();
       const parsedRoles = employees.map((employee) => ({
         ...employee,
@@ -32,6 +39,7 @@ export class EmployeeController {
 
       return res.status(200).json(parsedRoles);
     } catch (error) {
+      console.log("🚀 ~ EmployeeController ~ getAll ~ error:", error)
       throw error;
     }
   }
@@ -53,13 +61,19 @@ export class EmployeeController {
     }
   }
   @Post("/employee/create")
-  @UseBefore(validateRequest(employeeSchema))
+  @UseBefore(findOrCreateUser, validateRequest(employeeSchema), requireAuth())
   async createRole(@Req() req: Request, @Res() res: Response) {
     try {
-      console.log("🚀 ~ EmployeeController ~ createRole ~ req.body:", req.body);
-      const newRole = await EmployeeService.createEmployee(req.body);
-      res.status(201);
-      return res.status(200).json(newRole);
+      const auth = getAuth(req);
+      const userId = auth.userId ?? "";
+      const user = await UserService.getUserById(userId);
+      if (user) {
+        const newRole = await EmployeeService.createEmployee(req.body, user.id);
+        res.status(201);
+        return res.status(200).json(newRole);
+      } else {
+        res.status(403).json("Unauthorized");
+      }
     } catch (error) {
       throw error;
     }
@@ -73,6 +87,12 @@ export class EmployeeController {
     @Res() res: Response
   ) {
     try {
+      const employee = await EmployeeService.getEmploeeById(id);
+      const auth = getAuth(req);
+      const userId = auth.userId;
+      if (employee?.userId !== userId) {
+        return res.status(403).json({ msg: "Unauthorized" });
+      }
       const updatedEmployee = await EmployeeService.updateEmployee(
         id,
         req.body
@@ -89,6 +109,12 @@ export class EmployeeController {
     @Req() req: Request,
     @Res() res: Response
   ) {
+    const employee = await EmployeeService.getEmploeeById(id);
+    const auth = getAuth(req);
+    const userId = auth.userId;
+    if (employee?.userId !== userId) {
+      return res.status(403).json({ msg: "Unauthorized" });
+    }
     try {
       await EmployeeService.deleteRole(id);
       res.status(200).json({ msg: "Role deleted succesfully" });
